Rename the file common.c to utils.c
