+++ agent/src/ebpf/user/extended/extended.c
@@ -18,7 +18,7 @@
 #include <bcc/perf_reader.h>
 #include "../config.h"
 #include "../common_utils.h"
-#include "../common.h"
+#include "../utils.h"
 #include "../mem.h"
 #include "../log.h"
 #include "../types.h"
