+++ agent/src/ebpf/user/proc.c
@@ -40,7 +40,7 @@
 #include "config.h"
 #include "elf.h"
 #include "log.h"
-#include "common.h"
+#include "utils.h"
 #include "symbol.h"
 #include "proc.h"
 #include "tracer.h"
