+++ agent/src/ebpf/user/profile/perf_profiler.c
@@ -27,7 +27,7 @@
 #include <bcc/perf_reader.h>
 #include "../config.h"
 #include "../common_utils.h"
-#include "../common.h"
+#include "../utils.h"
 #include "../mem.h"
 #include "../log.h"
 #include "../types.h"
