+++ agent/src/ebpf/user/unwind_tracer.c
@@ -31,7 +31,7 @@
 #include "tracer.h"
 #include "socket.h"
 #include "proc.h"
-#include "common.h"
+#include "utils.h"
 #include "log.h"
 
 #include "load.h"
